+++ athina/__init__.py
@@ -3,14 +3,4 @@ from .guard.guard import guard
 from .guard.exception import AthinaGuardException
 from . import evals
 from . import keys
-from .runner.run import EvalRunner
-
-def run(evals, data):
-    """
-    A convenience wrapper to run evaluation suites.
-
-    :param evals: A list of evaluations to be run.
-    :param data: The dataset over which evaluations are run.
-    """
-    # Call the EvalRunner's run_suite method directly
-    return EvalRunner.run_suite(evals=evals, data=data)
+from .runner.run_wrapper import run
